don't log exception if file already doesn't exist

+++ verifier-repository/src/main/java/com/jetbrains/pluginverifier/misc/LanguageUtils.kt
@@ -22,7 +22,9 @@ fun Closeable.closeLogged() {
 
 fun File.deleteLogged(): Boolean {
   try {
-    FileUtils.forceDelete(this)
+    if (exists()) {
+      FileUtils.forceDelete(this)
+    }
     return true
   } catch(e: Exception) {
     LOG.error("Unable to delete $this", e)
